Move summary to stats module
