+++ pymc/__init__.py
@@ -6,6 +6,7 @@ from .math import *
 
 from .trace import *
 from .sample import *
+from .stats import summary
 from .step_methods import *
 from .tuning import *
